requirements.txt

openai==0.27.2
langchain==0.0.177
pinecone-client
python-dotenv
PyPDF2